requirement.txt
amqp==5.1.1
async-timeout==4.0.3
billiard==4.1.0
celery==5.3.4
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
dnspython==2.4.2
flower==2.0.1
humanize==4.8.0
kombu==5.3.2
prometheus-client==0.18.0
prompt-toolkit==3.0.39
pydantic==1.10.12
pymongo==4.5.0
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3.post1
redis==5.0.1
six==1.16.0
tornado==6.3.3
typing_extensions==4.8.0
tzdata==2023.3
vine==5.0.0
wcwidth==0.2.8
